pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ozzy"
version = "1.0.0"
description = "PIC simulation data analysis for the lazy and impatient"
authors = [
  "Mariana Moreira <mtrocadomoreira@gmail.com>",
]
license = "MIT"
readme = "README.md"
homepage = "https://mtrocadomoreira.github.io/ozzy/"
repository = "https://github.com/mtrocadomoreira/ozzy.git"
keywords = []
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "License :: OSI Approved :: MIT License",
  "Framework :: Hypothesis",
  "Framework :: Pytest",
  "Framework :: IPython",
  "Framework :: Jupyter",
  "Framework :: Jupyter :: JupyterLab",
  "Framework :: Matplotlib",
  "Framework :: MkDocs",
  "Intended Audience :: Science/Research",
  "Natural Language :: English",
  "Operating System :: MacOS",
  "Operating System :: Microsoft",
  "Operating System :: Unix",
  "Topic :: Scientific/Engineering :: Physics",
  "Topic :: Scientific/Engineering :: Visualization",
# Specify the Python versions you support here.
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = ">=3.10,<4"
xarray = "^2024.9.0"
tqdm = "^4.66.5"
numba = "^0.60.0"
h5py = "^3.12.1"
h5netcdf = "^1.4.0"
flox = "^0.9.13"
dask = "^2024.10.0"
dask-expr = "^1.1.16"
matplotlib = "^3.9.2"
cmcrameri = "^1.9"
seaborn = "^0.13.2"
ipython = "^8.28.0"
ipykernel = "^6.29.5"


[tool.poetry.group.optimize.dependencies]
numpy-groupies = "^0.11.2"
numbagg = "^0.8.2"

[tool.poetry.group.interactive.dependencies]
hvplot = "^0.11.1"
jupyter-bokeh = "^4.0.5"

[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.5.42"
mkdocstrings-python = "^1.12.2"
mkdocs-git-revision-date-localized-plugin = "^1.2.9"
mkdocs-glightbox = "^0.4.0"
black = "^24.10.0"
mkdocs-include-markdown-plugin = "^7.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
hypothesis = "^6.115.3"
pytest-mock = "^3.14.0"
python-semantic-release = "^9.12.0"
ruff = "^0.7.0"

[project]
name = "ozzy"
dynamic = ["version"]
dependencies = [
]

[tool.semantic_release]
version_variables = [
  "src/ozzy/__init__.py:__version__", 
]
version_toml = [
    "pyproject.toml:tool.poetry.version",
] 
branch = "main"
assets = []
build_command = "pip install poetry && poetry build"
build_command_env = []
commit_message = "{version}\n\nAutomatically generated by python-semantic-release"
commit_parser = "angular"
logging_use_named_masks = false
major_on_zero = true
allow_zero_version = true
no_git_verify = false
tag_format = "v{version}"

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease_token = "rc"
prerelease = false

[tool.semantic_release.changelog]
template_dir = "templates"
changelog_file = "CHANGELOG.md"
exclude_commit_patterns = ["test", "style", "chore"]

[tool.semantic_release.changelog.environment]
block_start_string = "{%"
block_end_string = "%}"
variable_start_string = "{{"
variable_end_string = "}}"
comment_start_string = "{#"
comment_end_string = "#}"
trim_blocks = false
lstrip_blocks = false
newline_sequence = "\n"
keep_trailing_newline = false
extensions = []
autoescape = true

[tool.semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "semantic-release <semantic-release>"

[tool.semantic_release.commit_parser_options]
allowed_tags = ["chore", "docs", "feat", "fix", "perf", "style", "refactor", "test"]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]
default_bump_level = 0

[tool.semantic_release.remote]
name = "origin"
type = "github"
ignore_token_for_push = false
insecure = false

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]

[tool.ruff]
